requirements.txt
streamlit
langchain
langchain-openai
langchain-experimental
pandas
faiss-cpu
python-dotenv
openai
tabulate
